moved uuid to its own file
