+++ lib/ex_debug_toolbar/breakpoint.ex
@@ -14,26 +14,8 @@ defmodule ExDebugToolbar.Breakpoint do
     :inserted_at
   ]
 
-  defmodule UUID do
-    defstruct [:request_id, :breakpoint_id]
-
-    def from_string(uuid) do
-      case String.split(uuid, "-") do
-        [request_id, breakpoint_id] ->
-          {:ok, %__MODULE__{request_id: request_id, breakpoint_id: String.to_integer(breakpoint_id)}}
-        _ -> {:error, "cannot parse uuid #{uuid}"}
-      end
-    end
-  end
-
   defdelegate start_iex(breakpoint_id, output_pid), to: Server, as: :start_link
   defdelegate send_input_to_iex(pid, input), to: Server, as: :send_input
   defdelegate stop_iex(pid), to: Server, as: :stop
   defdelegate code_snippet(env), to: Pry
-end
-
-defimpl String.Chars, for: ExDebugToolbar.Breakpoint.UUID do
-  def to_string(%{request_id: request_id, breakpoint_id: breakpoint_id}) do
-    "#{request_id}-#{breakpoint_id}"
-  end
 end
